ENH: DOS plotting from Questaal
Due to a design choice in Questaal, when LMF is run with DOS options a
total DOS will be written to *dos.ext* . LMDOS is then used to generate a
PDOS data in text form from the *moms.ext* file; but this will
overwrite the TDOS saved in *dos.ext* .

The Sumo way is to lay a PDOS over a TDOS where possible, so it will
look for a file called *tdos.ext* with an extension matching the dos
file. You need to move the *dos.ext* from LMF to this name before
running LMDOS.

LMF and LMDOS allow pretty much any combination of orbitals to be
written out and Sumo has no way of guessing which you chose. It also
already has syntax for combining orbitals that were decomposed by
[site, l, m], which happens to be the default output... So for now we
only support Questaal PDOS calculations which are fully site-l-m
decomposed.

As such the recommended procedure is:

1. Run you self-consistent `lmf` to converge density
2. Run lmf again with `--dos`, `--pdos` and `--quit=rho` options
   e.g.
   `lmf ctrl.ext --quit=rho --pdos~mode=2 --dos:npts=2001:window=-1,1`
   where "mode=2" is explicitly setting site-l-m decomposition (which
   should be the default anyway.)
3. MOVE _dos.ext_ TO _tdos.ext_
4. Run lmdos with the same `--dos`, `--pdos` settings as before
5. _dos.ext_ now contains PDOS data. It's quite large: you can
   compress it with gzip or bzip2 if you like.
6. sumo-dosplot --code questaal should detect the *dos.ext* or
   *dos.ext.gz* file, determine the extension and get other needed
   information from *tdos.ext* and *site.ext*.

Gaussian smearing is implemented. Bandgap / Fermi energy shenanigans
are currently not because I don't know of a standard place we can look
for eigenvalues.