ENH: Trailing '@' stripped from BS labels.
Another "secret labels" feature to aid with use of custom k-point
paths. Pymatgen handles high-symmetry points as a dictionary of labels
and positions. This means that it is prohibited to re-use the same
label for a different k-point, even if that position is actually
symmetry-equivalent or a periodic image in reciprocal space.

There are arguments for and against using such band structures, but
here we provide a workaround for those who wish to re-use labels for
non-equivalent points. By appending an '@' symbol to the label in
KPOINTS or syml.ext (i.e. 'X' becomes 'X@'), a new special point is
identified which will also display as 'X' in the plotted band
structure. What if you need another one? Just add more @ signs,
the code uses regular expressions to look for a block of @ signs at
the end of the label.