Refactor optics plotter to allow abs and loss as subplots
This is a more general solution than providing an optics plotter for
each spectrum, and allows the user to provide the property of interest
as an argument and optionally plot multiple subplots.

The amount of data is getting a bit complex here given that we can
also compare multiple data series and anisotropic directions. As it
hits the _plotter_ the main spectrum data structure is now of the form
e.g.::

    {'loss': [(energies, calc1_loss_spec),
              (energies, calc2_loss_spec),...],
     'absorption': [(energies, calc1_abs_spec),
                    (energies, calc2_loss_spec), ...]}

Where the spectral data is given as a vector of anisotropic data or
an array with (xx, yy, zz) columns. It is expected that the outer dict
will in fact be an OrderedDict, with the iteration order determining
the order of subplots.

User can specify global y ranges or give a sequence corresponding to
the plots, and 'auto' will be accepted to use automatic
scaling except for absorption spectra which take an automatic value of
1e5 cm-1 as this is the conventional scale for determining onset of
strong absorption. x axis limits are always global as the (vertical)
stack of plots has linked x axes for precise comparison.