Correct DOS alignment with CASTEP band structure
This is implemented differently from the VASP approach; we simply
read the offset from the existing band structure object and apply
that. It seems simpler and more consistent than what we are doing
for the VASP case; maybe this logic should be used for all codes?
(A few more lines would be needed to iterate through the PDOS terms.)