Castep: implement DOS alignment to VBM
Some VBM-detecting logic has been added for the electronic DOS from
CASTEP. This infrastructure seems a bit redundant; band structures do
this themselves in Pymatgen but there doesn't seem to be an equivalent
for the DOS. This might get messy if the implementations for DOS and
band structure disagree... Further revisions will probably be
appropriate as part of wider overhaul of alignment problems. Still,
it's a start and the Castep results are now aligned sensibly.