Castep: account for Efermi offset in DOS sampling
The sampling window for DOS construction is pre-shifted to account for
the energy zeroing, consistent with the "efermi" reported as part of
the DOS. Without this correction, values beyond emax of the
_unshifted_ data were being missed from the histogram binning, and
showing up as an empty region in the final output.

The histogram padding is not so necessary now, but has been tweaked to
adapt to the broadening width so that when an automatic range is set
the Gaussian tails will be included. (This affects the Python API, but
not the current CLI.)