Castep phonons: fix reading for LO-TO cases with extra entries
When Castep includes a LO-TO correction in the phonon data,
Gamma-point entries are duplicated with small deltas to capture
information in both directions. This means that the 'number of
wavevectors' reported in the header is NOT a reliable indicator of how
many blocks need to be read in the file. We add an extra scan over the
file to count the true number of blocks, and add this to our header
object as "n_entries".